add stuff

+++ src/app/manageArticle/page.tsx
@@ -6,6 +6,8 @@ import ClipLoader from "react-spinners/ClipLoader";
 import Auth from "@/components/Auth";
 import { getCategory } from "@/lib/features/categorySlice";
 import { useAppDispatch } from "@/lib/hooks";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 interface IManageArticleProps {}
 
@@ -27,9 +29,9 @@ const ManageArticle: React.FunctionComponent<IManageArticleProps> = (props) => {
   const category = useAppSelector((state) => state.categoryReducer);
   const dispatch = useAppDispatch();
   const [isLoading, setIsLoading] = React.useState(true);
+  const [deleteId, setDeleteId] = React.useState<string | null>(null);
 
   React.useEffect(() => {
-    // Simulasikan proses loading selama 2 detik
     const timer = setTimeout(() => {
       setIsLoading(false);
     }, 2000);
@@ -88,36 +90,51 @@ const ManageArticle: React.FunctionComponent<IManageArticleProps> = (props) => {
         );
         setEditId(null);
         setEditData(null);
+        toast.success("Article updated successfully!");
       } catch (error) {
         console.log(error);
+        toast.error("Failed to update article.");
       }
     }
   };
 
-  const handleDelete = async (articleId: string) => {
-    try {
-      const token = localStorage.getItem("user-token");
-      await axios.delete(
-        `https://blog-website-ashecone-25ef50f82ac6.herokuapp.com/articles/${articleId}`,
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      setArticles(articles.filter((article) => article.id !== articleId));
-    } catch (error) {
-      console.log(error);
-    }
+  const handleDeleteConfirm = (articleId: string) => {
+    setDeleteId(articleId);
   };
 
+  const handleDelete = async () => {
+    if (deleteId) {
+      try {
+        const token = localStorage.getItem("user-token");
+        await axios.delete(
+          `https://blog-website-ashecone-25ef50f82ac6.herokuapp.com/articles/${deleteId}`,
+          {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          }
+        );
+        setArticles(articles.filter((article) => article.id !== deleteId));
+        setDeleteId(null);
+        toast.success("Article deleted successfully!");
+      } catch (error) {
+        console.log(error);
+        toast.error("Failed to delete article.");
+      }
+    }
+  };
   const handleCancel = () => {
     setEditId(null);
     setEditData(null);
   };
 
+  const handleDeleteCancel = () => {
+    setDeleteId(null);
+  };
+
   return (
     <>
+      <ToastContainer />
       {isLoading ? (
         <div className="flex justify-center items-center h-screen">
           <ClipLoader size={150} color={"#123abc"} loading={true} />
@@ -135,28 +152,34 @@ const ManageArticle: React.FunctionComponent<IManageArticleProps> = (props) => {
             </div>
             <hr className="mb-6" />
             <div className="space-y-4">
-              {articles.map((article) => (
-                <div
-                  key={article.id}
-                  className="flex justify-between items-center"
-                >
-                  <span className="text-black">{article.title}</span>
-                  <div>
-                    <button
-                      className="text-white px-4 py-2 bg-blue-700 hover:bg-blue-400 hover:text-black mr-4"
-                      onClick={() => handleEdit(article)}
-                    >
-                      Edit
-                    </button>
-                    <button
-                      className="text-white px-4 py-2 bg-red-700 hover:bg-red-400 hover:text-black"
-                      onClick={() => handleDelete(article.id)}
-                    >
-                      Delete
-                    </button>
+              {articles.length === 0 ? (
+                <p className="text-black text-center">
+                  Not have any article, let's create one.
+                </p>
+              ) : (
+                articles.map((article) => (
+                  <div
+                    key={article.id}
+                    className="flex justify-between items-center"
+                  >
+                    <span className="text-black">{article.title}</span>
+                    <div>
+                      <button
+                        className="text-white px-4 py-2 bg-blue-700 hover:bg-blue-400 hover:text-black mr-4"
+                        onClick={() => handleEdit(article)}
+                      >
+                        Edit
+                      </button>
+                      <button
+                        className="text-white px-4 py-2 bg-red-700 hover:bg-red-400 hover:text-black"
+                        onClick={() => handleDeleteConfirm(article.id)}
+                      >
+                        Delete
+                      </button>
+                    </div>
                   </div>
-                </div>
-              ))}
+                ))
+              )}
             </div>
 
             {editId && (
@@ -263,6 +286,50 @@ const ManageArticle: React.FunctionComponent<IManageArticleProps> = (props) => {
                       onClick={handleSave}
                     >
                       Save
+                    </button>
+                  </div>
+                </div>
+              </div>
+            )}
+
+            {deleteId && (
+              <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
+                <div className="bg-white p-8 w-1/3 rounded-lg relative">
+                  <button
+                    className="absolute top-4 right-4 text-gray-500 hover:text-gray-700"
+                    onClick={handleDeleteCancel}
+                  >
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      className="h-6 w-6"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      stroke="currentColor"
+                    >
+                      <path
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                        strokeWidth={2}
+                        d="M6 18L18 6M6 6l12 12"
+                      />
+                    </svg>
+                  </button>
+                  <h2 className="text-2xl font-bold mb-4">Delete Article</h2>
+                  <p className="mb-4 text-black">
+                    Are you sure you want to delete this article?
+                  </p>
+                  <div className="flex justify-end mt-4">
+                    <button
+                      className="px-5 py-2 border rounded-md bg-gray-300 hover:bg-gray-500 hover:text-white text-black mr-4"
+                      onClick={handleDeleteCancel}
+                    >
+                      Cancel
+                    </button>
+                    <button
+                      className="px-5 py-2 border rounded-md bg-red-600 hover:bg-red-800 text-white"
+                      onClick={handleDelete}
+                    >
+                      Yes, Delete
                     </button>
                   </div>
                 </div>
